requirements.txt
pandas>=1.5.0
psycopg2-binary>=2.9.0
sqlalchemy>=1.4.0
boto3>=1.26.0
pyarrow>=10.0.0
dbt-core>=1.7.0
dbt-clickhouse>=1.7.0
